Extract TreeMenu component from Menu
